requirements.txt
# Core dependencies
PyMuPDF>=1.23,<2
pandas>=2.2,<3
numpy>=1.26,<2
tqdm>=4.66,<5
python-dotenv>=1,<2

# Google Gemini API
google-generativeai>=0.5,<1
google-api-core>=2,<3
google-auth>=2,<3
google-auth-oauthlib>=1,<2
googleapis-common-protos>=1,<2

# HTTP / Networking
requests>=2,<3
urllib3>=2,<3

# PDF generation
reportlab>=4,<5

# FastAPI server
fastapi>=0.109,<1
uvicorn[standard]>=0.27,<1
pydantic>=2,<3
